+++ minidump-stackwalk/tests/test-minidump-stackwalk.rs
@@ -182,6 +182,29 @@ fn test_json_symbols() {
 }
 
 #[test]
+fn test_json_symbols_disabled() {
+    // For a while this didn't parse right
+    let bin = env!("CARGO_BIN_EXE_minidump-stackwalk");
+    let output = Command::new(bin)
+        .arg("--json")
+        .arg("--pretty")
+        .arg("--no-symbolication")
+        .arg("../testdata/test.dmp")
+        .arg("../testdata/symbols/")
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .output()
+        .unwrap();
+
+    let stdout = String::from_utf8(output.stdout).unwrap();
+    let stderr = String::from_utf8(output.stderr).unwrap();
+
+    assert!(output.status.success());
+    insta::assert_snapshot!("json-pretty-symbols-disabled", stdout);
+    assert_eq!(stderr, "");
+}
+
+#[test]
 fn test_evil_json() {
     // For a while this didn't parse right
     let bin = env!("CARGO_BIN_EXE_minidump-stackwalk");